# apps/web/lib/gen/model/checkInResponse.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */
import type { XPEvent } from './xPEvent';
import type { CheckInResponseMilestoneReached } from './checkInResponseMilestoneReached';

export interface CheckInResponse {
  /** XP awarded from this check-in */
  xp_awarded: number;
  /** Total XP after check-in */
  new_total_xp: number;
  /** Level after check-in */
  new_level: number;
  /** Whether user leveled up */
  level_up: boolean;
  /**
   * Previous level (only if level_up is true)
   * @nullable
   */
  previous_level?: number | null;
  /** Whether streak was incremented */
  streak_incremented: boolean;
  /** Current streak after check-in */
  new_streak: number;
  /** Whether streak was reset due to inactivity */
  streak_broken: boolean;
  /** List of penalty events applied (if any) */
  inactivity_penalties?: XPEvent[];
  /**
   * Streak milestone achieved (if any)
   * @nullable
   */
  milestone_reached?: CheckInResponseMilestoneReached;
  message: string;
}