# apps/web/lib/gen/model/experienceHistoryResponse.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */
import type { XPEvent } from './xPEvent';

export interface ExperienceHistoryResponse {
  events: XPEvent[];
  /** Total number of XP events */
  total_count: number;
  /** Whether there are more events to fetch */
  has_more: boolean;
}