# apps/web/lib/gen/model/streakMilestone.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */

export interface StreakMilestone {
  /** Number of days for this milestone */
  days: number;
  /** XP reward for reaching this milestone */
  xp_reward: number;
  /** Whether user has achieved this milestone */
  achieved: boolean;
  /**
   * When milestone was achieved
   * @nullable
   */
  achieved_at?: string | null;
  /**
   * Days remaining to reach this milestone (if not achieved)
   * @nullable
   */
  days_remaining?: number | null;
}