# apps/web/lib/gen/model/streakMilestonesResponse.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */
import type { StreakMilestone } from './streakMilestone';

export interface StreakMilestonesResponse {
  /** Current streak days */
  current_streak: number;
  milestones: StreakMilestone[];
}