# apps/web/lib/gen/model/xPEventType.ts
/**
 * Generated by orval v8.0.0-rc.3 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */

export type XPEventType = typeof XPEventType[keyof typeof XPEventType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const XPEventType = {
  daily_login: 'daily_login',
  transaction: 'transaction',
  streak_milestone: 'streak_milestone',
  inactivity_penalty: 'inactivity_penalty',
  financial_goal: 'financial_goal',
} as const;