# packages/openapi/api.yaml
openapi: 3.0.3
info:
  title: Core API
  description: |-
    Core API
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/insights/month-snapshot:
    get:
      tags:
        - Insights
      summary: Get month snapshot
      description: Get detailed spending insights for a specific month including categories, weekly breakdown, and savings
      operationId: getMonthSnapshot
      parameters:
        - name: year
          in: query
          description: Year (e.g., 2025)
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
            example: 2025
        - name: month
          in: query
          description: Month (1-12)
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 12
      responses:
        '200':
          description: Month snapshot retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthSnapshot'
        '400':
          description: Invalid parameters
          $ref: '#/components/responses/ErrorResponse400'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '404':
          description: No data for month or no budget plan
          $ref: '#/components/responses/ErrorResponse404'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/insights/available-months:
    get:
      tags:
        - Insights
      summary: List available months
      description: Get list of months with transaction data, sorted by most recent first
      operationId: listAvailableMonths
      responses:
        '200':
          description: Available months retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableMonth'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'

  # Expense Categories
...
  /api/v1/experience/status:
    get:
      tags:
        - Experience
      summary: Get current experience status
      description: Get current level, XP, streak, and gamification stats
      operationId: getExperienceStatus
      responses:
        '200':
          description: Current experience status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperienceResponse'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/experience/check-in:
    post:
      tags:
        - Experience
      summary: Daily check-in
      description: Award daily login XP, update streak, and check for inactivity penalties
      operationId: checkIn
      responses:
        '200':
          description: Check-in result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/experience/history:
    get:
      tags:
        - Experience
      summary: Get XP transaction history
      description: Get paginated history of XP events
      operationId: getExperienceHistory
      parameters:
        - name: limit
          in: query
          description: Number of events to return
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of events to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: XP event history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperienceHistoryResponse'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'
  /api/v1/experience/streak-milestones:
    get:
      tags:
        - Experience
      summary: Get streak milestones and progress
      description: Get available streak milestones with achievement status
      operationId: getStreakMilestones
      responses:
        '200':
          description: Streak milestones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreakMilestonesResponse'
        '401':
          description: Unauthorized
          $ref: '#/components/responses/ErrorResponse401'
        '500':
          description: Internal server error
          $ref: '#/components/responses/ErrorResponse500'
components:
  schemas:
...
    GenericError:
      type: object
      description: Generic error response
      properties:
        status:
          type: integer
        error:
          type: string
          example: "INVALID_INPUT"
        message:
          type: string
          example: "Some error"
      required:
        - status
        - error
        - message
...
    SubCategoryBreakdown:
      type: object
      required:
        - id
        - total
        - percent
        - color
      properties:
        id:
          type: string
          example: "groceries"
        total:
          type: number
          format: float
          example: 160.5
        percent:
          type: number
          format: float
          example: 50.0
          description: Percentage of parent category total
        color:
          type: string
          example: "#f7b267"
    CategoryBreakdown:
      type: object
      required:
        - id
        - total
        - percent
        - items
        - color
      properties:
        id:
          type: string
          example: "essentials"
        total:
          type: number
          format: float
          example: 320.0
        percent:
          type: number
          format: float
          example: 42.0
          description: Percentage of total spending
        items:
          type: integer
          example: 18
          description: Number of transactions in this category
        color:
          type: string
          example: "#f59a3e"
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/SubCategoryBreakdown'
    WeeklySpendingPoint:
      type: object
      required:
        - week
        - label
        - total
      properties:
        week:
          type: integer
          minimum: 1
          maximum: 6
          example: 1
        label:
          type: string
          example: "Week 1"
        total:
          type: number
          format: float
          example: 180.5
    SavingsBreakdown:
      type: object
      required:
        - saved
        - invested
      properties:
        saved:
          type: number
          format: float
          example: 120.0
          description: Total from 'savings' category
        invested:
          type: number
          format: float
          example: 85.0
          description: Total from 'investments' category
        savedDelta:
          type: number
          format: float
          nullable: true
          example: 0.08
          description: Month-over-month percentage change for savings
        investedDelta:
          type: number
          format: float
          nullable: true
          example: 0.05
          description: Month-over-month percentage change for investments
    RecentTransactionSummary:
      type: object
      required:
        - amount
        - categoryId
        - date
      properties:
        amount:
          type: number
          format: float
          example: 42.5
        categoryId:
          type: string
          example: "essentials"
        subcategoryId:
          type: string
          nullable: true
          example: "groceries"
        date:
          type: string
          format: date-time
          example: "2025-12-04T00:00:00Z"
    MonthSnapshot:
      type: object
      required:
        - key
        - label
        - currentDate
        - loggedDays
        - totalDays
        - totalSpent
        - budget
        - lastMonthDelta
        - categories
        - savings
        - weekly
        - transactions
      properties:
        key:
          type: string
          example: "dec-2025"
          description: Unique month identifier
        label:
          type: string
          example: "December 2025"
        currentDate:
          type: string
          format: date-time
          example: "2025-12-06T00:00:00Z"
        loggedDays:
          type: integer
          example: 12
          description: Number of days with expense transactions
        totalDays:
          type: integer
          example: 30
          description: Total days in the month
        totalSpent:
          type: number
          format: float
          example: 756.0
          description: Sum of all expense transactions
        budget:
          type: number
          format: float
          example: 1200.0
          description: From budget_plan.expected_income
        lastMonthDelta:
          type: number
          format: float
          example: 0.12
          description: Percentage change from previous month (-1.0 to +infinity)
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryBreakdown'
        savings:
          $ref: '#/components/schemas/SavingsBreakdown'
        weekly:
          type: array
          items:
            $ref: '#/components/schemas/WeeklySpendingPoint'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/RecentTransactionSummary'
          description: Last 5 expense transactions
    AvailableMonth:
      type: object
      required:
        - key
        - label
        - year
        - month
      properties:
        key:
          type: string
          example: "dec-2025"
        label:
          type: string
          example: "December 2025"
        year:
          type: integer
          example: 2025
        month:
          type: integer
          minimum: 1
          maximum: 12
          example: 12

    # Category Schemas
...
    EvolutionStage:
      type: string
      enum:
        - Baby
        - Young
        - Adult
        - Prime
        - Legendary
      example: "Baby"
    XPEventType:
      type: string
      enum:
        - daily_login
        - transaction
        - streak_milestone
        - inactivity_penalty
        - financial_goal
      example: "daily_login"
    ExperienceResponse:
      type: object
      required:
        - user_id
        - current_level
        - current_xp
        - xp_for_next_level
        - total_xp_for_current_level
        - evolution_stage
        - current_streak
        - longest_streak
        - transactions_today_count
        - transactions_daily_limit
      properties:
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        current_level:
          type: integer
          example: 15
          description: Current player level
        current_xp:
          type: integer
          example: 1200
          description: Total accumulated XP
        xp_for_next_level:
          type: integer
          example: 150
          description: XP needed to reach next level
        total_xp_for_current_level:
          type: integer
          example: 1050
          description: Total XP needed to reach current level (cumulative)
        evolution_stage:
          $ref: '#/components/schemas/EvolutionStage'
        current_streak:
          type: integer
          example: 14
          description: Current consecutive days streak
        longest_streak:
          type: integer
          example: 30
          description: Longest streak ever achieved
        last_login_date:
          type: string
          format: date
          nullable: true
          example: "2025-01-20"
          description: Last check-in date (server-side)
        transactions_today_count:
          type: integer
          example: 3
          description: Number of transactions logged today (for XP cap)
        transactions_daily_limit:
          type: integer
          example: 5
          description: Maximum transactions that count toward XP per day
    CheckInResponse:
      type: object
      required:
        - xp_awarded
        - new_total_xp
        - new_level
        - level_up
        - streak_incremented
        - new_streak
        - streak_broken
        - message
      properties:
        xp_awarded:
          type: integer
          example: 15
          description: XP awarded from this check-in
        new_total_xp:
          type: integer
          example: 1215
          description: Total XP after check-in
        new_level:
          type: integer
          example: 15
          description: Level after check-in
        level_up:
          type: boolean
          example: false
          description: Whether user leveled up
        previous_level:
          type: integer
          nullable: true
          example: 14
          description: Previous level (only if level_up is true)
        streak_incremented:
          type: boolean
          example: true
          description: Whether streak was incremented
        new_streak:
          type: integer
          example: 15
          description: Current streak after check-in
        streak_broken:
          type: boolean
          example: false
          description: Whether streak was reset due to inactivity
        inactivity_penalties:
          type: array
          items:
            $ref: '#/components/schemas/XPEvent'
          description: List of penalty events applied (if any)
        milestone_reached:
          type: object
          nullable: true
          description: Streak milestone achieved (if any)
          properties:
            days:
              type: integer
              example: 7
            xp_reward:
              type: integer
              example: 50
        message:
          type: string
          example: "Welcome back! +15 XP"
    XPEvent:
      type: object
      required:
        - id
        - xp_amount
        - event_type
        - description
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        xp_amount:
          type: integer
          example: 15
          description: XP amount (can be positive or negative)
        event_type:
          $ref: '#/components/schemas/XPEventType'
        description:
          type: string
          example: "Daily check-in"
        created_at:
          type: string
          format: date-time
          example: "2025-01-20T10:00:00Z"
    ExperienceHistoryResponse:
      type: object
      required:
        - events
        - total_count
        - has_more
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/XPEvent'
        total_count:
          type: integer
          example: 145
          description: Total number of XP events
        has_more:
          type: boolean
          example: true
          description: Whether there are more events to fetch
    StreakMilestone:
      type: object
      required:
        - days
        - xp_reward
        - achieved
      properties:
        days:
          type: integer
          example: 7
          description: Number of days for this milestone
        xp_reward:
          type: integer
          example: 50
          description: XP reward for reaching this milestone
        achieved:
          type: boolean
          example: true
          description: Whether user has achieved this milestone
        achieved_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-15T10:00:00Z"
          description: When milestone was achieved
        days_remaining:
          type: integer
          nullable: true
          example: 3
          description: Days remaining to reach this milestone (if not achieved)
    StreakMilestonesResponse:
      type: object
      required:
        - current_streak
        - milestones
      properties:
        current_streak:
          type: integer
          example: 14
          description: Current streak days
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/StreakMilestone'
  responses:
...
    ErrorResponse400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
    ErrorResponse500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'